meta-ampere: Update ampere_fanctrl.sh scripts
This commit updates ampere_fanctrl.sh scripts in Ampere's platforms to
change their behaviors as follows:

  1. Don't allow to enable fan control when it's already running, or
disable when it's already off.
  2. Don't allow to set fan speed when automatic fan control is enabled.
  3. Support a setspeed force command for internal scripts to set fan
speeds regardless of the automatic fan control status. [1]
  4. Have systaxes, variable names and error logs reinforced to get rid
of shellcheck issue and for structure enhancement.
  5. Have systemd service masking/unmasking scheme for setting fan
control status replaced by utilizing another service to conflict with
phosphor-fan services and prevent them from starting when this service
is active. The new service is named "ampere-fanctrl-manual.service" and
is added to systemd under ampere-utils recipe.
"ampere-fanctrl-manual.service" is started when fan control is disabled
and stopped when it's enabled.
  6. For Mt.Jefferson, have fan control status not depend on
phosphor-pid-control.service if host power status is Off.

Tested:

Test case 1:

1. Get fan control status: $ ampere_fanctrl.sh getstatus
=> If the output is "Thermal Control operational status: Enabled"
1.1 Enable fan control status: $ ampere_fanctrl.sh setstatus 0
=> Expected output is "Fan control status is already enabled"

2. If the fan control status query command output is
"Thermal Control operational status: Disabled"
2.1 Disable fan control status: $ ampere_fanctrl.sh setstatus 1
=> Expected output is "Fan control status is already disabled"

Test case 2:

1. Get fan control status: $ ampere_fanctrl.sh getstatus
=> If the output is "Thermal Control operational status: Enabled"
1.1 Set fan speed for each fan $ ampere_fanctrl.sh setspeed $fan $pwm
=> Expected output is "Setting speeds is not allowed when fan control is
enabled."

2. If the fan control status query command output is
"Thermal Control operational status: Disabled"
2.1 Set fan speed for each fan $ ampere_fanctrl.sh setspeed $fan $pwm
=> Fan speed can be successfully set
2.3 Check fan speed again to compare $ ampere_fanctrl.sh getspeed $fan
=> Output match the desired fan speed previously set

Test case 3:

1. When fan control status is Enabled
1.1 Force setting fan speeds: $ampere_fanctrl.sh setspeed force all 100
=> Fan speed can be successfully set

Test case 4:

BMC is on, Host is on.

1. Disable fan control with "ipmitool raw 0x3c 0x03 0x01"
2. Turn off host power => fans preserve the previous speeds
3. Turn on host power => phosphor-fan services are not started
4. Enable fan control with "ipmitool raw 0x3c 0x03 0x00"
=> phosphor-fan services are started

Test case 5:

BMC is on, Host is on.

1. Turn off host power => fans run at 10% PWM
2. Disable fan control => fans preserve the previous speeds (10%)
3. Enable fan control => phosphor-fan services are started, but
   phosphor-pid-control service is not started, fans still run at 10%.
4. Get fan control status => Enabled (when host is Off, does not care
   about phosphor-pid-control)
5. Turn on host power => all fan services starts

+++ meta-ampere/meta-jefferson/recipes-ampere/platform/ampere-utils/ampere_fanctrl.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-# shellcheck disable=SC2004
 
 fan_hwmon_num=$(ls /sys/bus/i2c/drivers/adt7462/8-005c/hwmon)
-fan_hwmon_path="/sys/class/hwmon/$fan_hwmon_num"
+fan_hwmon_path="/sys/class/hwmon"
+
+DISABLED_STAT="Disabled"
+ENABLED_STAT="Enabled"
+ACTIVE_STR="active"
+MANUAL_CONTROL_SERV="ampere-fanctrl-manual.service"
 
 phosphor_fan_service=("phosphor-fan-control@0.service"
                       "phosphor-fan-monitor@0.service"
@@ -10,27 +14,43 @@ phosphor_fan_service=("phosphor-fan-control@0.service"
                       "phosphor-pid-control.service")
 
 declare -A fan_id_list
+# The index in $input_list and $pwm_list for each FAN
 fan_list=(0 1 2 3)
+# Mapping from real FAN index to the index in $fan_list
 fan_id_list=([0]=0 [1]=1 [2]=2 [3]=3)
+# The index of fan*_input file in sysfs for each fan
 input_list=(1 3 5 7)
+# The index of pwm* file in sysfs for each fan
 pwm_list=(1 2 3 4)
 
+function print_fan_indices(){
+    indices=("${!fan_id_list[@]}")
+    echo "${indices[@]}"
+}
+
 function stop_phosphor_fan_services() {
+    status=$(getstatus)
+    if [[ "$status" == *"$DISABLED_STAT" ]]; then
+        echo "Fan control status is already disabled"
+        exit 0
+    fi
+
+    systemctl start $MANUAL_CONTROL_SERV
+
     for service in "${phosphor_fan_service[@]}"
     do
         systemctl stop "$service"
-        busctl call org.freedesktop.systemd1 /org/freedesktop/systemd1 org.freedesktop.systemd1.Manager MaskUnitFiles asbb 1 "$service" true true
     done
-    systemctl daemon-reload
 }
 
 function start_phosphor_fan_services() {
-    for service in "${phosphor_fan_service[@]}"
-    do
-        busctl call org.freedesktop.systemd1 /org/freedesktop/systemd1 org.freedesktop.systemd1.Manager UnmaskUnitFiles asb 1 "$service" true
-    done
+    status=$(getstatus)
+    if [[ "$status" == *"$ENABLED_STAT" ]]; then
+        echo "Fan control status is already enabled"
+        exit 0
+    fi
 
-    systemctl daemon-reload
+    systemctl stop $MANUAL_CONTROL_SERV
 
     for service in "${phosphor_fan_service[@]}"
     do
@@ -43,13 +63,17 @@ function start_phosphor_fan_services() {
 }
 
 function getstatus() {
-    status="Enabled"
+    status="$ENABLED_STAT"
     ret=0
     for service in "${phosphor_fan_service[@]}"
     do
-        service_stt=$(systemctl is-active "$service" | grep inactive)
-        if [ -n "$service_stt" ]; then
-            status="Disabled"
+        service_stat=$(systemctl is-active "$service")
+        if [ "$service_stat" != "$ACTIVE_STR" ]; then
+            if [ "$service" == "phosphor-pid-control.service" ] &&
+            [ "$(obmcutil chassisstate | awk -F. '{print $NF}')" == 'Off' ]; then
+                continue
+            fi
+            status="$DISABLED_STAT"
             ret=1
             break
         fi
@@ -76,18 +100,23 @@ function setspeed() {
 
     if [ -n "$id" ]
     then
+        if ! { [[ "$pwm_val" =~ ^[0-9]+$ ]] && [ "$pwm_val" -ge 0 ] && [ "$pwm_val" -le 100 ]; }; then
+            echo "Error: Invalid target PWM value - ${pwm_val}. The valid range is [0-100]"
+            exit 1
+        fi
+
         # Get pwm sysfs file of the fan
         fan_pwm=pwm${pwm_list[$id]}
         # Scale PWM to 255, adding 50 for rounding.
-        pwm_scaled=$(((($pwm_val * 255) + 50) / 100))
+        pwm_scaled=$((((pwm_val * 255) + 50) / 100))
 
-        if ! (echo "$pwm_scaled" > "${fan_hwmon_path}/${fan_pwm}");
+        if ! (echo "$pwm_scaled" > "${fan_hwmon_path}/${fan_hwmon_num}/${fan_pwm}");
         then
             echo "Error: Set Fan PWM Duty Cycle failed!"
             exit 255
         fi
     else
-        echo "Error: Fan $fan_index doesn't exist"
+        echo "Error: Fan $fan_index doesn't exist. Valid indices are: $(print_fan_indices)"
         exit 1
     fi
 }
@@ -101,14 +130,14 @@ function getspeed() {
         fan_pwm=pwm${pwm_list[$id]}
 
         # Get fan speed and pwm from hwmon
-        tach_reading=$(cat "${fan_hwmon_path}/${fan_input}")
+        tach_reading=$(cat "${fan_hwmon_path}/${fan_hwmon_num}/${fan_input}")
         ret=$?
         if [ $ret != 0 ]; then
             echo "Error: Get Fan Speed failed!"
             exit 1
         fi
 
-        pwm_reading=$(cat "${fan_hwmon_path}/$fan_pwm")
+        pwm_reading=$(cat "${fan_hwmon_path}/${fan_hwmon_num}/${fan_pwm}")
         ret=$?
         if [ $ret != 0 ]; then
             echo "Error: Get Fan PWM Duty Cycle failed!"
@@ -120,7 +149,7 @@ function getspeed() {
 
         echo "FAN${fan_index}, PWM: $pwm_reading, Duty cycle: $pwm_reading_scaled%, Speed(RPM): $tach_reading"
     else
-        echo "Error: Fan $fan_index doesn't exist"
+        echo "Error: Fan $fan_index doesn't exist. Valid indices are: $(print_fan_indices)"
         exit 1
     fi
 }
@@ -128,7 +157,7 @@ function getspeed() {
 # Usage of this utility
 function usage() {
     echo "Usage:"
-    echo "  ampere_fanctrl.sh [getstatus] [setstatus <0:enable|1:disable>] [setspeed all/<fan> <duty>] [getspeed <fan>]"
+    echo "  ampere_fanctrl.sh [getstatus] [setstatus <0:enable|1:disable>] [setspeed all/<fan> <duty>] [getspeed all/<fan>]"
     echo "  fan: 0-3"
     echo "  duty: 1-100"
 }
@@ -138,6 +167,16 @@ if [ "$1" == "getstatus" ]; then
 elif [ "$1" == "setstatus" ] && [ -n "$2" ]; then
     setstatus "$2"
 elif [ "$1" == "setspeed" ]; then
+    if [ "$2" == "force" ]; then
+        # Shift to the command argument position
+        shift
+    else
+        status=$(getstatus)
+        if [[ "$status" == *"$ENABLED_STAT" ]]; then
+            echo "Setting speeds is not allowed when fan control is enabled."
+            exit 1
+        fi
+    fi
     if [ "$2" == "all" ] && [ -n "$3" ]; then
         for fan in "${fan_list[@]}"
         do
